map product intro to cms component
