+++ projects/storefrontlib/src/cms-components/cms-lib.module.ts
@@ -29,14 +29,15 @@ import {
   NavigationModule,
   SearchBoxModule,
 } from './navigation/index';
+import { OrderConfirmationModule } from './order-confirmation/index';
 import {
   ProductCarouselModule,
+  ProductIntroModule,
   ProductListModule,
   ProductReferencesModule,
   ProductTabsModule,
 } from './product/index';
 import { ProductImagesModule } from './product/product-images/product-images.module';
-import { OrderConfirmationModule } from './order-confirmation/index';
 import { ProductSummaryModule } from './product/product-summary/product-summary.module';
 
 @NgModule({
@@ -71,6 +72,7 @@ import { ProductSummaryModule } from './product/product-summary/product-summary.
     // StoreFinderModule,
     ProductImagesModule,
     ProductSummaryModule,
+    ProductIntroModule,
     CheckoutComponentModule,
     ForgotPasswordModule,
     ResetPasswordModule,
