+++ projects/storefrontlib/src/cms-components/product/index.ts
@@ -1,11 +1,13 @@
+export * from './carousel/index';
 export * from './carousel/product-carousel/product-carousel.component';
 export * from './carousel/product-carousel/product-carousel.module';
 export * from './carousel/product-references/product-references.component';
 export * from './carousel/product-references/product-references.module';
-export * from './product-summary/product-summary.component';
-export * from './product-summary/product-summary.module';
-export * from './carousel/index';
 export * from './current-product.service';
+export * from './product-intro/product-intro.component';
+export * from './product-intro/product-intro.module';
 export * from './product-list/index';
 export * from './product-outlets.model';
+export * from './product-summary/product-summary.component';
+export * from './product-summary/product-summary.module';
 export * from './product-tabs/index';
