+++ projects/storefrontlib/src/cms-components/product/product-intro/product-intro.component.html
@@ -1,0 +1,18 @@
+<ng-container *ngIf="(product$ | async) as product">
+  <div class="rating">
+    <cx-star-rating
+      [rating]="product?.averageRating"
+      [disabled]="true"
+    ></cx-star-rating>
+    <div class="count">({{ product?.numberOfReviews }})</div>
+    <a
+      class="btn-link"
+      *ngIf="(reviewsTabAvailable | async)"
+      (click)="showReviews()"
+      >{{ 'productSummary.showReviews' | cxTranslate }}</a
+    >
+  </div>
+  <div class="code">
+    {{ 'productSummary.id' | cxTranslate }} {{ product?.code }}
+  </div>
+</ng-container>
