+++ projects/storefrontlib/src/cms-components/product/product-summary/product-summary.component.html
@@ -1,27 +1,5 @@
 <ng-container *ngIf="(product$ | async) as product">
   <ng-template
-    [cxOutlet]="outlets.INTRO"
-    [cxOutletContext]="{ product: product }"
-  >
-    <div class="rating">
-      <cx-star-rating
-        [rating]="product?.averageRating"
-        [disabled]="true"
-      ></cx-star-rating>
-      <div class="count">({{ product?.numberOfReviews }})</div>
-      <a
-        class="btn-link"
-        *ngIf="(reviewsTabAvailable | async)"
-        (click)="showReviews()"
-        >{{ 'productSummary.showReviews' | cxTranslate }}</a
-      >
-    </div>
-    <div class="code">
-      {{ 'productSummary.id' | cxTranslate }} {{ product?.code }}
-    </div>
-  </ng-template>
-
-  <ng-template
     [cxOutlet]="outlets.PRICE"
     [cxOutletContext]="{ product: product }"
   >
