+++ projects/storefrontstyles/scss/layout/page-templates/_product-detail.scss
@@ -6,25 +6,25 @@
       @extend .container;
       margin-bottom: 40px;
 
-      @include media-breakpoint-up(lg) {
-        display: grid;
-        grid-template-columns: 1fr 1fr;
-        grid-template-rows: 1fr 1fr;
-        grid-column-gap: 20px;
-        grid-row-gap: 0px;
-        cx-product-images {
-          grid-area: 1 / 1 / 3 / 2;
-        }
+      //   @include media-breakpoint-up(lg) {
+      //     display: grid;
+      //     grid-template-columns: 1fr 1fr;
+      //     grid-template-rows: 1fr 1fr;
+      //     grid-column-gap: 20px;
+      //     grid-row-gap: 0px;
+      //     cx-product-images {
+      //       grid-area: 1 / 1 / 3 / 2;
+      //     }
 
-        cx-product-summary {
-          grid-area: 1 / 2 / 2 / 3;
-        }
+      //     cx-product-summary {
+      //       grid-area: 1 / 2 / 2 / 3;
+      //     }
 
-        cx-add-to-cart {
-          grid-area: 2 / 2 / 3 / 3;
-          padding: 0 20px 20px 20px;
-        }
-      }
+      //     cx-add-to-cart {
+      //       grid-area: 2 / 2 / 3 / 3;
+      //       padding: 0 20px 20px 20px;
+      //     }
+      //   }
     }
     &.AddToCart {
       max-width: 50%;
